fixed registration authorization flow
