+++ nurego_lib/nurego-authentication.js
@@ -6,8 +6,10 @@ module.exports = (function() {
     //Get token by grant_type=password
     getToken: function(params){
       console.log("i am here 4\n",params);
+      console.log("typeof params",typeof params);
       return new Promise(function(resolve,reject){
-  console.log("i am here 4",params);
+        var auth={};
+  console.log("i am here 4a",params);
         var basicAuth= new Buffer(config.uaa.client_id+':'+config.uaa.client_secret).toString('base64');
         params.grant_type = "password";
         var options = {
@@ -25,43 +27,24 @@ module.exports = (function() {
             reject({ error: error});
           }else if(response.statusCode != 200 ){
               console.log("i am here 5\n",{statusCode: response.statusCode, error: body });
-            reject({statusCode: response.statusCode, error: body });
+              auth = {
+                isLoggedIn: false,
+                token: null,
+                errors: body
+              };
+            reject({statusCode: response.statusCode, auth: auth });
           }else{
               console.log("i am here 6");
-            resolve({statusCode: response.statusCode, token: body });
+              auth = {
+                isLoggedIn: true,
+                token: body,
+                errors: null
+              };
+            resolve({statusCode: response.statusCode, auth: auth });
           }
         }
-        request(options, callback);
-      });
-    },
-    authorize: function(params){
-      console.log("i am here 3\n",params);
-      return new Promise(function(resolve,reject){
-        //Authorize the cutomer automatically
-        var auth={};
-        var _this=this;
-          console.log("i am here 3\n",params);
-        _this.getToken(params)
-        .then((data) => {
-            console.log("i am here 7");
-          auth = {
-            isLoggedIn: true,
-            token: data.token,
-            errors: null
-          };
-          resolve({auth: auth});
-        })
-        .catch((data) => {
-            console.log("i am here 8");
-          auth = {
-            isLoggedIn: false,
-            token: null,
-            errors: data.error
-          };
-          reject({auth: auth});
-        });
+      request(options, callback);
       });
     }
   }; //return
-})
-();
+})();
